Use contact fields for segmentation survey trigger
Previously it used keywords, which was problematic if the user sent `1` instead if `Hell yeah!`, but also problematic with features that need to be added. Now it uses contact fields from RapidPro.

It also updates the rapidpro mock for the segment survey to be a bit easier to assert on